requirements.txt
flask
ultralytics
google-cloud-vision
openai
